requirements.txt
Django==1.10.6
django-nested-inline==0.3.7
